pyproject.toml
[tool.poetry]
name = "hexlet-flask"
version = "0.1.0"
description = ""
authors = ["Safar Galimzyanov <safargalimzyanov@protonmail.com>"]
readme = "README.md"
packages = [{include = "hexlet_flask"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "3.0"
email-validator = "^1.1.3"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
